requirements.txt
asgiref==3.10.0
certifi==2025.11.12
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
django-modeltranslation==0.19.12
gunicorn==23.0.0
sentry-sdk==2.47.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.1
